fix(api,agent): Make the common FNN internal route-target explicitly configurable
While working on HFA, we realized that some values GNI expects aren't going to work in newer DCs because the ASN+VNI will exceed the 6-byte limit for a community/RT.  

This MR makes the values explicitly/separately configurable.  They're being left optional so that they both have to be explicitly configured in the Carbide config.  If unset, there will be no double-tagging of internal tenant routes.  There's a good chance the "common ASN" value will always be the same as as the "datacenter ASN" value in the future, but being explicit is ok.

See merge request nvmetal/carbide!4346