feat: Record partition keys associated with each GUID
With this change, IBFabricMonitor will try to detect all the pkeys associated with a particular port/GUID by queriyng the UFM API.

It will then use this data to create an enhanced version of the `MachineIbInterfaceStatusObservation` which stores the associated pkeys per Interface on each Machine.

The data can be viewed as part of the `ib_status` fields on the Machine.

The data has 2 purposes:

1. It helps debugging. We can now detect whether UFM thinks a particular interface is associated with certain pkeys or not without performing manual UFM queries.
2. It will form the base for continuous reconciliation (https://jirasw.nvidia.com/browse/FORGE-6424)

Example for a Machines IB status:

```json
{"ib_interfaces":[{"guid":"a088c203002c216a","lid":16,"fabric_id":"default","associated_pkeys":["0x13"]}],"observed_at":"2025-06-06T19:47:16.597282585Z"}
```

See merge request nvmetal/carbide!4361