fix: Fix fetching partition information in IbFabricMonitor
The `get_ib_networks` method for reading GUID data didn't work in IbFabricMonitor due to a mixture of the UFM bug and the code assuming that QoS data is always returned. To avoid the problem, we read the partition details by looping over all partitions after we read the partition keys. That is slower, but should still work.

See merge request nvmetal/carbide!4362