fix: Show IB Partition Keys (pkeys) in hex format
The hexadecimal notation is the canconcial format of pkeys. So far we showed the pkey in decimal format to users in `IBPartitionStatus::pkey`. This change changes that to hex format.

Research so far showed that no tenant actively looks at the pkey value, therefore this change should not be disruptive to any user.

See merge request nvmetal/carbide!4375