feat: Add new IB metrics to check whether IB configuration is synced against UFM
Adds 2 new metrics:
- `forge_ib_monitor_machines_with_missing_pkeys_count`: The amount of machines where at least one port is not assigned to the expected pkey on UFM
- `forge_ib_monitor_machines_with_unexpected_pkeys_count`: The amount of machines where at least one port is assigned to an unexpected pkey on UFM

If any of these metrics show a value different then 0, it indicates a desync between Carbide and UFM. The state-machine is expected to fix that desync on short term. But if the value is > 0 for a longer time manual fixes on the UFM configuration will be required as long as there is no automatic reconcilation.

Note: the metrics have a limitation at the moment where they don't precisely check whether the pkey associated with a GUID is the expected pkey. They only check whether there is any pkey associated with the GUID. This will be fixed in the next release.

See merge request nvmetal/carbide!4377