fix sync writes in async code; make file output available to all commands (not...
This moves the file creation for the `-o` option earlier in main so every command has the ability to use it.

fixes existing commands that use the `-o` option to use async writers to avoid blocking calls in async code. While not a big deal for the CLI, copying the broken code could be dangerous.  (this came up during a previous MR)

Adds macros for handling async + format

* `async_write!` for writing formatted output to a tokio::io::AsyncWrite object
* `async_writeln!` for writing formatted output with a newline to a tokio::io::AsyncWrite object
* `async_write_table_as_csv! `for writing a prettytable table as csv to a tokio::io::AsyncWrite object

See merge request nvmetal/carbide!4378