MLT: Rebuild MLT image on every merge
The machine lifecycle test image is built on the forge-admin-cli image. However, we don't rebuild it regularly so we're using an old version of the admin-cli.

This is an issue when we need access to the latest fixes.

Changing it so the MLT image rebuilds after every MR is merged to trunk, along with all the other container builds.

See merge request nvmetal/carbide!4396