ssh-console: Potential fix for defunct ipmitool processes
ssh-console in dev3 is not getting data from some ipmi consoles, and the
process table is showing it's not properly wait()'ing on ipmitool:

```
UID          PID    PPID  C STIME TTY          TIME CMD
root           1       0  4 Aug08 ?        02:55:39 /opt/carbide/ssh-console run -c /etc/forge/ssh-console/config.toml
root         260       1  0 Aug08 ?        00:00:00 [ipmitool] <defunct>
```

What *should* be happening is that if ipmitool dies, the pty master fd
should start failing read requests, which should cause the message loop
to break and for us to finally kill/wait on the pid (and kill -9'ing
ipmitool locally shows this does happen), but that's not happening in
dev3.

So to attempt fixing this, have the message loop also check the child
pid with `.wait()` in the tokio select, so that if the process dies we
explicitly exit (rather than relying on the fd read failure.)

Another potential issue is if the `.kill()` call is stuck forever (it's
supposed to await the process exit status), so do that in the background
just in case.

Also, log the output of ipmitool as part of reporting its failure, so
that we get a better idea of what's going on.

See merge request nvmetal/carbide!4416