ssh-console: Use custom errors instead of eyre
Using eyre with color errors is nice in development to reduce
boilerplate and get nice backtrace-esque things, but it clogs the logs
in loki with a bunch of ansi escapes, preventing logfmt from working. It
also keeps outputting the "Run with RUST_BACKTRACE=1" notice on every
error, which is overly verbose, not to mention that multi-line errors
are just not that great from a reporting perspective.

Use thiserror instead, and try to have a unique error type for every
function and error case, and nest them accordingly, so that we always
have annotated errors for exactly what went wrong.

See merge request nvmetal/carbide!4430