feat(agent): Enforce stricter ARP/neighbor learning
We learned recently during an incident (1160 in rootly) that the default configurations of both neighmgrd and ifupdown2 on a bridge subinterface allow learning neighbor (read: ARP) entries that are outside of the IP prefix associated with an interface.

If the IP address in question is one associated with an anycast VIP (where we want each host and DPU to use the L3 routing table and not the L2 neighbor table), this will result in churning on the MAC address associated with the entry, which in turn results in TCP connections to the VIP suddenly being remapped to a different tenant instance, breaking the connection with a TCP RST from the instance the packet has landed on.

I suspect this also may have allowed a tenant to hijack an IP belonging to another tenant, or an IP entirely outside the site.

The fix for this requires two config tweaks inside the HBN container:
1. The neighmgrd config file (`/etc/cumulus/neighmgr.conf`) must contain `subnet_checks=1` in the `[snooper]` section.
2. An ifupdown2 policy file (here we use `/etc/network/ifupdown2/policy.d/forge-arp-accept.json`) must contain `{"address":{"module_globals":{"l3_intf_arp_accept":"0"}}}`.

This MR applies these changes on agent startup or when the HBN container ID changes.

See merge request nvmetal/carbide!4433