ssh-console: Support log rotation
Per FORGE-6734, the existing legacy ssh-console is growing to large log
sizes. The logs are available in loki as the system of record, so we
don't actually need to store them in our container for very long. (We
don't want to just send them straight to loki though, as there are other
use cases, like for checking hardware health via the console logs, which
also need to read the log files.)

So perform basic rotation, with a generous (configurable) default of
10MB limit before rotation and 3 rotated logs. The otelcol-contrib
collector supports log rotation out of the box so it will work fine.

There are related changes in nvmetal/forged!4154, but they're separable
from this change, so these can be merged in either order (otelcol still
works fine with log rotation today, that MR is more about avoiding
potentially missing lines during a rotation period.)

See merge request nvmetal/carbide!4436