feat: Record associated_partition_ids for each of a Machines IB interfaces
In addition to record `associated_pkeys` which each IB interface, this change now also records `associated_partition_ids` by looking up the Partition IDs that map to the pkey.

That will allow the Machine state handler to determine whether an IB config is fully applied to a Host without having to load the IB partition details.

The `associated_partition_ids` is shown on the web-ui together with `associated_pkeys`.

It is possible that a Machine is associated with a pkey that is not known to Forge. In this case, the length of the `associated_partition_ids` and `associated_pkeys` arrays will be different. If this is the case, then the Machines IB sync status would be treated as "not synced".

A new metric `forge_ib_monitor_machines_with_unknown_pkeys_count` is emitted that counts whether any machines are associated with pkeys that are unknown to Forge.

See merge request nvmetal/carbide!4449