ssh-console-rs: Reconnect faster by waiting for BMC to be up first
I'm using exponential backoff to wait longer and longer if the
connection fails, but there are cases when DPU's go offline for a
while, then when they come back up, we want to connect right away. So
before attempting a connection, wait indefinitely for the BMC to be "up"
(ping for ipmi-based, TCP half-open for SSH-based) before attempting a
connection. That way the backoff only happens for "real" failures (like
incorrect password, connection in use, etc.)

See merge request nvmetal/carbide!4446