ssh-console: Fix for DPU consoles not returning any data
For DPU's we can skip the part where we detect the BMC prompt and
activate the serial console, but we *can't* skip the part where we
allocate a PTY and request a shell. Move the early return statement down
a few lines to make sure we still do those parts.

Currently no tests expose this bug, because machine-a-tron is not
launching a BMC mock for DPUs. This will be fixed in a followup MR, the
fix is what's important right now.

See merge request nvmetal/carbide!4461