fix: switch from generic type param -> impl trait arg
Switching a few cases of simple generic type parameters
over to impl trait arguments -- as the docs say, this is
purely just syntactic sugar, unless we wanted to be able
to explicitly specify a generic type for `T` at the call
site. In our case, I don't think we really care about
that here, so I'm proposing a little sugar.

There's more examples of this in the codebase, but I wanted
to start here to see what the general feel was.

See merge request nvmetal/carbide!4471