ssh-console: Emit messages when connected/disconnected from BMC's
This was something that worked early on when logging was in charge of
the connection/retry logic, but when that was refactored out, the logs
only say "connected" when the actual process starts, so basically only
ever once.

Fix this by broadcasting messages to the frontends when
disconnection/reconnections happen, which will cause the messages to be
written to the logs as well.

See merge request nvmetal/carbide!4481