ssh-console: Fix crash in exponential backoff time
`next_retry_backoff` was originally written with the assumption that the
previous duration would always either be either Duration::ZERO or some
amount greater than or equal to RETRY_BASE_DURATION, but a recent change
broke this assumption by deducting time spent waiting for the port to be
open from the total sleep time.

This can result in the "upper" part of the random range being smaller
than BASE, which causes a panic (random_range panics on an empty range.)

Main fix is to ensure the random range is between prev..upper and not
base..upper, since upper may be smaller than base.

But also, write tests to assert on reconnect behavior, which involves
adjusting the timeouts so the tests don't take too long. Confirmed that
these tests catch the bug.

See merge request nvmetal/carbide!4483