feat: API, DB and State Machine changes to support SPX
Test methodology used on my dev env machine for testing these changes:

1) When the ingested host was in READY state, use grpcurl to create and associate 8 SPX NICs with the machine.
2) Create Instance.
3) When Machine is in WaitingForNetworkConfig state, use grpcurl to update the SPX interface network observation status, mimicking
   what would happen when we get a SetVNI Ack.
4) Instance gets created and machine is in Assigned state.
5) Release instance.
6) Once again, when machine is in WaitingForNetworkReconfig state, use grpcurl to update the SPX interface network observation status.
7) Machine moves back to READY state.
8) Tested force_delete

The create/delete/update API interfaces are purely for debugging and testing purposes and might be removed later on.
Unit tests written for API calls that are intended for the final product.

The next set of changes will be to integrate this code with the EMQ client.

See merge request nvmetal/carbide!4343