formalize mlxconfig variable management with mlxconfig variable registry
We're going to have to start managing configuration on tens of
thousands of DPAs pretty soon, and we already manage configuration on
thousands of DPUs.

Instead of managing configuration on them via exec'ing `mlxconfig`
commands in `bash` and hoping for the best, I'm putting together a
mechanism for formalizing how we manage configuration on Mellanox
devices.

I did a demo of this a number of weeks back on configuring "profiles"
for devices and syncing expected values to them, and this is the
groundwork for that.

Originally, my demo had me hard-coding variable definitions, but
one of my TODOs was to be able to define different registry schemas
depending on hardware type. The idea is not all cards support the
same variables, so we should be able to have different variable
registries with constraints per device type, part number, and/or
even firmware version.

The idea is if we want to apply a variable setting to a specific
Mellanox device, it should be:
- A variable we have allowed ourselves to manage.
- A value we know to be valid for that variable.

Using this, we can manage configuration on:
-  `forge-dpu-agent` (for DPUs)
- `scout` (for DPAs)
- AND, we can also manage configuration via `carbide-api` -- having known
variables + value specs can allow us to push validated config
down to approved devices safely.

The next part of this is the profile definitions and management,
but that will come in subsequent MRs, including the `mlxconfig`
runner itself.

I have various `README.md` files for this, including:
- `dev/mlxconfig_embedded/README.md` -- An example CLI to play around.
- `common/mlxconfig/README.md` -- An overall readme of the system.
- `common/mlxconfig/registry/README.md` -- Discusses the registry builder.
- `common/mlxconfig/variables/README.md` -- The type definitions for the registry.

See merge request nvmetal/carbide!4511