perf: Faster /admin/managed-host page
This brings the time taken to render /admin/managed-host from 22 seconds
down to 0.6 seconds on my machine, on a database with 800 managed hosts.

This is the result of a bit of an odyssey after trying find a way to minimize
how much information we get from the database when indexing managed hosts (since
we only need a tiny subset of the information.) You can find the dead-end on
writing an entirely separate set of types/queries just for indexing managed
hosts [here][dead-end].

Turns out that the query wasn't really the bottleneck, but it was all the extra
work we were doing inside ManagedHostMetadata::lookup_from_api, which consists
of getting an entire site exploration report, finding connected devices (for
things like switch ports info), finding network devices, etc. We don't need any
of this when showing a non-detailed list of managed hosts. (We do need it in the
managed-host.json view, as there's integration with SRE tools that use the
connected devices info, but that can be potentially sped up in future work.)

So to fix this, just have managed_host::show_html convert the managed hosts to
HTML rows directly, rather than laundering them through ManagedHostMetadata
first.

[dead-end]: https://gitlab-master.nvidia.com/nvmetal/carbide/-/compare/trunk...ksimon%2Fdeadend-faster-managed-host-list

See merge request nvmetal/carbide!4536