add forge_alerts_suppressed_count metric
This MR adds a `forge_alerts_suppressed_count` metric, similar to the existing `forge_hosts_unhealthy_by_classification_count` metric minus the classification label (the value of the time series is 1 only if the SuppressAlerts classification is present) and with an added `machine_id` label. This will be useful for suppressing external alerts by matching them to hosts with the SuppressAlerts classification.

See merge request nvmetal/carbide!4415