chore: Fix libredfish version
Cargo is pulling in 0.29.72, which has breaking changes. Nobody has
intentionally upgraded libredfish to 0.29.72 yet (the last MR was !4538
which bumped to 0.29.71 and merged yesterday) but for some reason cargo
thinks it's ok to move right on up to 0.29.72, and the build is broken
for me.

Pin it harder at 0.29.71 using the `=` spec to prevent inadvertent
updates, until we're ready to adopt 0.29.72 for real.

See merge request nvmetal/carbide!4554