chore: Unify to a single MachineId type
## Description

One of the issues we've had with tonic is that the types it spits out are less strict than the types we want: MachineId is the most glaring example here. The gRPC type is just a string, but our "rich" MachineId type has constraints on what ID's are valid (prefix, length, machine type, etc.)

This has resulted in a lot of places where we're simply passing the plain-old-string gRPC type around out of convenience, so that we don't have to go through a fallible conversion. But this is a bad habit: It means that random places are doing conversions late (when they bother to care about the type), when we want to do the conversion as close to the gRPC boundary as possible.

Luckily, tonic lets you pick custom types to use for messages. So just tell it directly to use the gRPC type instead of generating one.

If an incoming machine_id is invalid, older clients will get an error that looks like:

```
The API call to the Forge API server returned status: Internal, message: "failed to decode Protobuf message: MachinesByIdsRequest.machine_ids: Invalid machine id: fake_machine_id"
```

Newer clients won't even be able to construct requests using invalid machine ID's in the first place.

Next up after this is to do similar work for all the other ID types.


## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
<!-- How was this tested? Check all that apply -->
- [X] Unit tests added/updated
- [X] Integration tests added/updated  
- [X] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

There should be no functional changes here. It doesn't break gRPC compatibility, because the MachineId type has a custom prost::Message impl which mimics how the previous rpc::MachineId struct was being serialized. I checked with using an old forge_admin_cli against this branch deployed to my devenv and it works fine.

See merge request nvmetal/carbide!4555