chore: Use type-safe UUID wrappers directly in RPC messages
## Description
As a follow-on from !4555, this moves various common.UUID fields to their corresponding rust UUID wrappers. For instance, RPC methods which pass an Instance ID as a UUID will now show up in rust code as `id: InstanceId` rather than `id: common::Uuid`.

This avoids needing to deal with fallible conversions to our internal types, instead the protobuf message itself will fail to decode.

Similarly to !4555, the messages are wire-compatible with the old common.UUID values, so this should all be backwards compatible.

## Type of Change
<!-- Check one that best describes this MR -->
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [X] Unit tests added/updated
- [X] Integration tests added/updated  
- [X] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

This is broken into a commit for each type that is being changed. Technically these are all separable... they could each be individual MR's if we wanted to.

See merge request nvmetal/carbide!4566