feat: Prevent Instance configuration changes when instance is terminating
## Description

If an instances configuration is changed while it is terminating - it could lead to unexpected side effects - such as the instance being stuck in termination due to waiting for not the most recent configurations to be applied.
To prevent that, this change prevents an instances configuration from being changed after it has been marked as deleted.

## Type of Change
- [ ] **Add** - New feature or capability
- [x] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [x] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4558