adding mlxconfig-runner; pulling together registry + variables
## Description

This adds the `mlxconfig-runner` crate, pulling together the work
that was done to create `mlxconfig-variables`, which made it so we
could define the `mlxconfig-registry`.

Now you can do things like this, which will be integrated into
`scout` for working with DPA settings, and the `forge-dpu-agent`
for working with DPU settings (**which we can then report up via
metrics to make sure all devices are in sync with what the expected
settings are**):
```
// Leverage an existing MlxConfigVariable from a registry.
let dpa_options = registry.get("dpa_options");
let dscp = dpa_options.get("MULTIPATH_DSCP");
let new_value = dscp.with("DSCP_0)";
runner.set(new_value);

// Just toss strings at it and let the crate validate, where
// it will pull the MlxConfigVariable from the registry
// behind the scenes and attempt to convert the value into
// the correct type (in this case it actually turns it into
// an enum option.
runner.set([
    ("MULTIPATH_DSCP", "DSCP_0"),
    ("ROCE_ADAPTIVE_ROUTING_EN", true),
]);
```

With these in place, we can now explicitly define and work with
known `mlxconfig` variables, which have type safety within the
system, and let us control them in a way that's better than just
firing off `mlxconfig` commands in shell scripts.

The idea is we define managed variables in the `registry`, where
we set the type, any options (e.g. if it's an array, it has a max
size, and if it's an enum, it has valid options), and that is
all backed by the `variables` crate.

From there, we can work with the `mlxconfig-runner` to define
type-backed variable values that we want to apply to a card. Behind
the scenes, it *does* execute `mlxconfig` (it has to), but it
allows us to be very explicit in what we do.

Feature work coming next is:
- Defining profiles that we can "sync" to devices, leveraging
known variables from the registry.  We can say "sync this config
to this card", and it will do it.
- Being able to manage `mlxconfig` based variables on DPUs (and
DPAs) via the `forge-admin-cli` -- if we need to tweak something
on the fly, we can do it without needing to log into DPUs, or hosts
in admin mode for DPAs.

## Type of Change
This is new feature work that is still not in use anywhere yet, and is part of the groundwork being laid down for managing DPA configuration profiles, and it turns out will also be useful for managing DPU settings as well.

- [x] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [x] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)
JIRA: [FORGE-6256](https://jirasw.nvidia.com/browse/FORGE-6256), [FORGE-6102](https://jirasw.nvidia.com/browse/FORGE-6102)

## Breaking Changes
- [ ] This MR contains breaking changes

<!-- If checked above, describe the breaking changes and migration steps -->

## Testing
<!-- How was this tested? Check all that apply -->
- [x] Unit tests added/updated
- [x] Integration tests added/updated  
- [x] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

Also updated the `mlxconfig-embedded` CLI reference example (in `dev/mlxconfig-embedded`).

See merge request nvmetal/carbide!4574