support for managing mlxconfig profiles against devices
## Description
This introduces the `mlxconfig-profile` crate, which pulls together
the `mlxconfig-{registry,runner,variables}` crates to define registry
backed profiles, which we can then compare [changes against] and
sync [changes to] devices.

Profiles can be programatically created, or they can be placed into
config file and distributed.

The idea is we can define a profile like so:
```
name: bf3-spx-on
registry_name: mlx_generic
description: Config settings to apply to a Bluefield-3 for SPX RA.
config:
  INTERNAL_CPU_OFFLOAD_ENGINE: ENABLED
  ROCE_ADAPTIVE_ROUTING_EN: true
  USER_PROGRAMMABLE_CC: true
  TX_SCHEDULER_LOCALITY_MODE: STATIC_MODE
  MULTIPATH_DSCP: DEVICE_DEFAULT
  ROCE_RTT_RESP_DSCP_P1: 48
  ROCE_RTT_RESP_DSCP_MODE_P1: FIXED_VALUE
```

And then do something like this:
```
let profile = MlxConfigProfile::from_yaml(<that-yaml>)?;
if profile.compare(<device-id>)? {
    profile.sync(<device-id>)?
}
```

In other words, we're able to look if there are changes
between the config on the card and what we specified in
the profile, and then sync those changes.

Moving forward, we can either hard-code profiles into code,
OR, put them into `forged` and deploy them with a site,
and then export metrics on cards, their profile, and if they
are out of sync or not.

In general, `scout` (for DPAs), and `forge-dpu-agent` (for DPUs)
would take care of the compare + sync management, so no one
should have to do anything. BUT, if a card remained out of
sync with the desired state of its profile, we would have
awareness into it.

## Type of Change
This is still introducing features and groundwork for DPA work, but just so happens to also be very useful for DPUs. None of this code is in use yet, and is going to start to be integrated soon -- almost time!

- [x] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [x] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)
<!-- If applicable, provide JIRA or/and NVBugs ticket. If none, the MR itself will be referenced in changelog -->
JIRA: [FORGE-6256](https://jirasw.nvidia.com/browse/FORGE-6256), [FORGE-6102](https://jirasw.nvidia.com/browse/FORGE-6102)

## Breaking Changes
- [ ] This MR contains breaking changes

<!-- If checked above, describe the breaking changes and migration steps -->

## Testing
<!-- How was this tested? Check all that apply -->
- [x] Unit tests added/updated
- [ ] Integration tests added/updated  
- [x] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4577