Add ssh-console-rs and ssh-console permission to call GetMachine
## Description
For some reason ssh-console-rs works just fine without this rule in
every env I've tested it in (and yes, rbac is enabled in each of the
relevant envs, including local-dev and dev3, dev4, etc.) But it's
finally biting us in ytl, so fix it.

## Type of Change
<!-- Check one that best describes this MR -->
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [X] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
<!-- How was this tested? Check all that apply -->
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes
This is untested but I cannot reproduce the original bug either, so I can't really test it.

See merge request nvmetal/carbide!4578