perf: Rerun IbFabricMonitor task earlier in case of changes
## Description

Reruns the IbFabricMonitor after only 1s in case changes to the fabric have been applied in a previous iteration.
This speeds up the time an instance switches between states, since it no longer is blocked for 1 Minute until the next regular iteration of the Fabric Monitor occurs and the outcome of the changes are actually recorded in the database.

## Type of Change

- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [x] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)

## Breaking Changes
- [ ] This MR contains breaking changes


## Testing

- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4581