Fix clippy-flow build issues
## Description
For some reason we had an issue with `cargo make clippy-flow` but not `cargo make pre-commit-verify-workspace`, even though the latter calls the former.

The issue is that if the `cli` feature is active in the `measured-boot` crate, it requires the `cli` feature of the `rpc` crate too, but it wasn't activating it. This only caused an issue if you try to build measured-boot by itself though, otherwise the `rpc/cli` feature is already enabled via other dependencies.

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4587