enhancing mlxconfig mgmt w/ device info integration + filtering + reporting
## Description

https://jirasw.nvidia.com/browse/FORGE-6256

This adds what should be the final piece to the collection
of `mlxconfig-*` crates, and I'll probably end up merging all
of it into a single crate, but wanted to do it crate by crate
just to make the MRs a little more digestable.

Ultimately what this does is makes it easy to query for a
single device via `mlxfwmanager` (or *ALL*) devices, using
the XML interface to collect + parse information. One use
is for sending `MlxDeviceReport` updates for doing DPA, as
well as DPU, management. We can do this within `scout` or
the `forge-dpu-agent`:
```
let device_report = MlxDeviceReport::new().collect()?;
let request = PublishMlxDeviceReportRequest {
  report: Some(device_report.into()),
};
let response = api_client.PublishMlxDeviceReport(request)?;
tracing::debug(response);
```

This ALSO introduces device filtering alongside it, allowing
me to rip out the constraints stuff I had been baking in. I
decided I liked this more, and that it was kind of a duplication
having both constraints and filtering. With constraints, it was
technically `mlxconfig` backed, so I was limited to what fields
we could "constrain" on. With the new filtering, it's `mlxfwmanager`
backed, so we can set filters on anything visible via that. The
idea is to ensure we won't apply profile configuration to incompatible
devices. For example, we can do something like:
```
let device = device::load("01:00.0")?;
if !registry.filters.matches(&device) {
  tracing::warn("device {device} not allowed to sync per registry filters");
}
```

In more practical terms, we can create a "DPA" filter to just
get DPAs, a "DPU" filter to just get DPUs, etc.

Finally, this also introduces an `MlxDeviceReport` struct, which
has an accompanying protobuf and trait implementations, which
will be used for sending full device reports to an API endpoint
for use with DPA config + version management, and will extend
to also being used for DPUs as well.

The `mlxconfig-embedded` reference CLI has been updated with
some commands to mess around with the interface (for manual
testing), as well as updated unit tests.

## Type of Change
None of this code is in use yet by any features, and is just continuing to lay the framework for building on for the DPA work (and I'll also fold DPUs into this eventually too).

- [x] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [x] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)
<!-- If applicable, provide JIRA or/and NVBugs ticket. If none, the MR itself will be referenced in changelog -->
JIRA: [FORGE-6256](https://jirasw.nvidia.com/browse/FORGE-6256), [FORGE-6102](https://jirasw.nvidia.com/browse/FORGE-6102)

## Breaking Changes
- [ ] This MR contains breaking changes

<!-- If checked above, describe the breaking changes and migration steps -->

## Testing
<!-- How was this tested? Check all that apply -->
- [x] Unit tests added/updated
- [x] Integration tests added/updated  
- [x] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4593