feat: FORGE-6437: added limit on max reports
Added a forge-admin-cli command to trim tables with historical data. This particular MR is just about Measured Boot tables. The number of entries/rows to keep is configurable, but it is advised to keep it at 250, as this seems to be the convention.
This CronJob MR https://gitlab-master.nvidia.com/nvmetal/forged/-/merge_requests/4255 will be the one typically invoking the trim-table command with the --keep-entries set to 250.

See merge request nvmetal/carbide!4366